scx_mitosis: Fix runnable task stall
This fixes an error I created in PR #2978, which lead to mitosis getting kicked out due to runnable task stalls.

When we update cell dsq vtime in stopping(), we need to update the cell dsq that corresponds to the CPU, not the task.

These are often the same thing, but there are special cases where a task belonging to one cell is pinned to a CPU that belongs to another cell, e.g. a pinned kworker thread.