ci: drop vm for cargo test
Currently we run the equivalent of `cargo test`, `cargo nextest`, in a
VM as root. This was not a good idea as it allows people to write tests
that rely on this, and some distros when they package want to run the
cargo tests and don't have the ability to run them in a VM. This means
we should constrain our `cargo test` to run without root.

Also, this VM is causing problems. This is partially because the CI
runner setup is a bit weird at the minute, but give it's also not a good
idea let's drop it again and run the tests directly on the host as the
unprivileged user. If we want VMs in Cargo tests we need to adjust the
test to launch a VM somehow.

Test plan:
- CI