scx_lavd: Revise the logic for shrinking slice at ops.tick().
When there are pinned tasks waiting to run on this CPU, we need to shrink
the current task's time slice. If the task's running time so far is already
beyond the upper bound (pinned_slide_ns vs. sys_stat.slice), shrink it to zero
right now for immediate preemption. Also, we canceled the slice boost.

Signed-off-by: Changwoo Min <changwoo@igalia.com>