scx_lavd: Tracking the number of pinned tasks on a CPU precisely.
Previously, we didn't correctly track the number of pinned tasks when a
task changed its pin location. Now, we track the pinned CPU ID in task_ctx
(pinned_cpu_id), which is set to -ENOENT when the task is not pinned or not
runnable. Then increment and decrement the count when the task becomes
enqueued and dequeued, respectively.

Signed-off-by: Changwoo Min <changwoo@igalia.com>