scx_lavd: Arrange neighbors in a circular sorted order.
Previously, the neighbor list of a compute domain was numerically
sorted. When load balancing, we started from a random position in the
sorted neighbor list. This approach is suboptimal because a task from
one domain can be migrated to any other domain; there is no
proximity-based task migration, which further hurts cache locality.

Let's mitigate this problem by traversing neighbor domains in circular
distance order. Suppose there are eight domains. We can place them on a
circle.

      0
   7     1
 6         2
   5     3
      4

For domain 0, its neighbor list is [1, 7, 2, 6, 3, 5, 4]; for domain 1,
it is [0, 2, 3, 7, 4, 6, 5]; for domain 7, it is [0, 6, 1, 5, 2, 4, 3].
Now, 0, 1, 7 are in proximity in stealing order.

Signed-off-by: Changwoo Min <changwoo@igalia.com>