scx_cosmos: Introduce automatic GPU / NUMA affinity
Introduce the new options --gpu / --gpu-util-threshold to enable
automatic GPU affinity on multi-GPU systems. When enabled, the scheduler
will attempt to keep tasks using a GPU on the NUMA node associated with
that device.

In the future, this mechanism can be generalized to support affinity for
other device types as well, evolving from GPU-specific affinity into a
broader device-aware scheduling policy.

Signed-off-by: Andrea Righi <arighi@nvidia.com>