scx_lavd: Use the utility function to count the # of queued tasks for a waker's CPU.
In the case of synchronous wake-up, the waker's CPU is not considered
an idle CPU in the idle bitmask because it is executing the wake-up path.
So, we need to check the number of queued tasks. Let's use the utility
function, nr_queued_on_cpu(), for this. Plus, nr_queued_on_cpu() also counts
the queued tasks on local DSQ, making idle checking more precise.

Signed-off-by: Changwoo Min <changwoo@igalia.com>